Code Cleanups
