+++ Files/BaseLayout.cs
@@ -153,7 +153,7 @@ namespace Files
             }
             App.CurrentInstance.NavigationToolbar.CanRefresh = true;
             IsItemSelected = false;
-            AssociatedViewModel.EmptyTextState.IsVisible = Visibility.Collapsed;
+            AssociatedViewModel.IsFolderEmptyTextDisplayed = false;
             App.CurrentInstance.ViewModel.WorkingDirectory = parameters;
 
             // pathRoot will be empty on recycle bin path
@@ -181,10 +181,6 @@ namespace Files
             base.OnNavigatingFrom(e);
             // Remove item jumping handler
             Window.Current.CoreWindow.CharacterReceived -= Page_CharacterReceived;
-            if (App.CurrentInstance.ViewModel._fileQueryResult != null)
-            {
-                App.CurrentInstance.ViewModel._fileQueryResult.ContentsChanged -= App.CurrentInstance.ViewModel.FileContentsChanged;
-            }
             App.AppSettings.LayoutModeChangeRequested -= AppSettings_LayoutModeChangeRequested;
         }
 
@@ -289,7 +285,7 @@ namespace Files
                 AssociatedInteractions = App.CurrentInstance.InteractionOperations;
                 if (App.CurrentInstance == null)
                 {
-                    App.CurrentInstance = ItemViewModel.GetCurrentSelectedTabInstance<ModernShellPage>();
+                    App.CurrentInstance = InstanceTabsView.GetCurrentSelectedTabInstance<ModernShellPage>();
                 }
             }
         }
