+++ Files/Files.csproj
@@ -163,6 +163,7 @@
     <Compile Include="Helpers\DispatcherHelper.cs" />
     <Compile Include="Helpers\ItemsDataTemplateSelector.cs" />
     <Compile Include="Helpers\NativeDirectoryChangesHelper.cs" />
+    <Compile Include="Helpers\NativeFindStorageItemHelper.cs" />
     <Compile Include="Helpers\NaturalStringComparer.cs" />
     <Compile Include="Helpers\PackageHelper.cs" />
     <Compile Include="Helpers\StringExtensions.cs" />
@@ -220,7 +221,6 @@
     <Compile Include="Views\InstanceTabsView.xaml.cs">
       <DependentUpon>InstanceTabsView.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Interacts\EmptyFolderTextState.cs" />
     <Compile Include="Interacts\Interaction.cs" />
     <Compile Include="Interacts\ItemInteractions.cs" />
     <Compile Include="IShellPage.cs" />
