+++ Files/UserControls/LayoutModes/GridViewBrowser.xaml
@@ -576,8 +576,8 @@
             Canvas.ZIndex="0"
             Text="This folder is empty."
             TextWrapping="Wrap"
-            Visibility="{x:Bind AssociatedViewModel.EmptyTextState.IsVisible, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
-
+            x:Load="{x:Bind AssociatedViewModel.IsFolderEmptyTextDisplayed, Mode=OneWay}" />
+        
         <muxc:TeachingTip
             x:Name="FileNameTeachingTip"
             x:Uid="FileNameTeachingTip"
